requirements.txt
asgiref==3.8.1
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
Django==5.0.7
django-cors-headers==4.4.0
django-crontab==0.7.1
django-dotenv==1.4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.53.1
idna==3.7
joblib==1.4.2
kiwisolver==1.4.5
krwordrank==1.0.3
load-dotenv==0.1.0
matplotlib==3.8.0
meson==1.5.1
ninja==1.11.1.1
numpy==1.26.4
packaging==24.1
pillow==10.4.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reportlab==4.2.2
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.2
wheel==0.43.0
wordcloud==1.9.3
